# openapi_schema.json
openapi: 3.0.3
info:
  title: BitTorrent Tracker API
  version: 1.0.0
  description: A comprehensive BitTorrent tracker backend with user management, credit
    system, and advanced analytics.
paths:
  /api/auth/register/:
    post:
      operationId: auth_register_create
      description: Register a new user account with invite code validation.
      summary: User Registration
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              SuccessfulRegistration:
                value:
                  username: johndoe
                  email: john@example.com
                  password: securepassword123
                  password_confirm: securepassword123
                  invite_code: ABC123DEF
                summary: Successful Registration
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
  /api/auth/login/:
    post:
      operationId: auth_login_create
      description: Authenticate user and return JWT tokens.
      summary: User Login
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
            examples:
              LoginRequest:
                value:
                  username: johndoe
                  password: securepassword123
                summary: Login Request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
...
  /api/auth/invite/create/:
    post:
      operationId: auth_invite_create_create
      description: ایجاد کد دعوت جدید (فقط برای ادمین‌ها)
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      - {}
      responses:
        '200':
          description: No response body
  /api/auth/invite/generate/:
    post:
      operationId: auth_invite_generate_create
      description: ایجاد کد دعوت توسط کاربر عادی (نیاز به کلاس member یا بالاتر)
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          description: No response body
  /api/auth/invite/my-codes/:
    get:
      operationId: auth_invite_my_codes_retrieve
      description: نمایش کدهای دعوت ایجاد شده توسط کاربر
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          description: No response body
  /api/user/profile/:
    get:
      operationId: user_profile_retrieve
      description: Get and update authenticated user profile information.
      summary: User Profile
      tags:
      - User Management
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: ''
        '401':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
    put:
      operationId: user_profile_update
      description: Get and update authenticated user profile information.
      summary: User Profile
      tags:
      - User Management
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: ''
        '401':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
    patch:
      operationId: user_profile_partial_update
      description: Get and update authenticated user profile information.
      summary: User Profile
      tags:
      - User Management
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUserProfileUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUserProfileUpdateRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: ''
        '401':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
...
  /api/user/tokens/:
    get:
      operationId: user_tokens_list
      description: نمای لیست و ایجاد توکن‌های احراز هویت
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - user
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuthTokenList'
          description: ''
    post:
      operationId: user_tokens_create
      description: نمای لیست و ایجاد توکن‌های احراز هویت
      tags:
      - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthTokenCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthTokenCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AuthTokenCreateRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      - Bearer:
          type: apiKey
          name: Authorization
          in: header
          description: 'JWT Bearer token authentication. Format: Bearer {token}'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenCreate'
          description: ''
...
components:
  schemas:
...
    AuthToken:
      type: object
      description: Serializer برای توکن‌های احراز هویت
      properties:
        id:
          type: integer
          readOnly: true
        token:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
        is_active:
          type: boolean
      required:
      - created_at
      - expires_at
      - id
      - token
    AuthTokenCreate:
      type: object
      description: Serializer برای ایجاد توکن جدید
      properties:
        ip_bound:
          type: string
          nullable: true
    AuthTokenCreateRequest:
      type: object
      description: Serializer برای ایجاد توکن جدید
      properties:
        ip_bound:
          type: string
          nullable: true
          minLength: 1
...
    PaginatedAuthTokenList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/AuthToken'
...
    PatchedUserProfileUpdateRequest:
      type: object
      description: Serializer for updating user profile including file uploads
      properties:
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        profile_picture:
          type: string
          format: binary
          nullable: true
          description: User profile picture
...
    UserClassEnum:
      enum:
      - newbie
      - member
      - trusted
      - elite
      type: string
      description: |-
        * `newbie` - Newbie
        * `member` - Member
        * `trusted` - Trusted
        * `elite` - Elite
    UserLoginRequest:
      type: object
      description: Serializer برای لاگین
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - password
      - username
...
    UserProfile:
      type: object
      description: Serializer برای پروفایل کاربر
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        user_class:
          allOf:
          - $ref: '#/components/schemas/UserClassEnum'
          readOnly: true
        total_credit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        locked_credit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        available_credit:
          type: string
          readOnly: true
        lifetime_upload:
          type: integer
          readOnly: true
        lifetime_download:
          type: integer
          readOnly: true
        ratio:
          type: string
          readOnly: true
        download_multiplier:
          type: string
          readOnly: true
        max_torrents:
          type: string
          readOnly: true
        is_banned:
          type: boolean
          readOnly: true
        date_joined:
          type: string
          format: date-time
          readOnly: true
        last_login:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        profile_picture:
          type: string
          format: uri
          nullable: true
          description: User profile picture
      required:
      - available_credit
      - date_joined
      - download_multiplier
      - id
      - is_banned
      - last_login
      - lifetime_download
      - lifetime_upload
      - locked_credit
      - max_torrents
      - ratio
      - total_credit
      - user_class
      - username
    UserProfileUpdateRequest:
      type: object
      description: Serializer for updating user profile including file uploads
      properties:
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        profile_picture:
          type: string
          format: binary
          nullable: true
          description: User profile picture
      required:
      - username
    UserRegistrationRequest:
      type: object
      description: Serializer برای ثبت‌نام کاربر
      properties:
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 8
        password_confirm:
          type: string
          writeOnly: true
          minLength: 1
        invite_code:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - invite_code
      - password
      - password_confirm
      - username
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...